+++ schedulesync/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "0.1.2"  # noqa: C0114
+__version__ = "0.1.3"  # noqa: C0114