+++ schedulesync/main.py
@@ -2,7 +2,7 @@ from fastapi import FastAPI
 from starlette.requests import Request
 from starlette.responses import Response
 
-from schedulesync.core.models.db import SessionLocal
+from schedulesync.core.auth.db import async_session_maker
 from schedulesync.routes import routes
 
 tags_metadata = [
@@ -27,7 +27,7 @@ app = FastAPI(openapi_tags=tags_metadata)
 async def db_session_middleware(request: Request, call_next):
     response = Response("Internal server error", status_code=500)
     try:
-        request.state.db = SessionLocal()
+        request.state.db = async_session_maker()
         response = await call_next(request)
     finally:
         request.state.db.close()